pyproject.toml
[project]
name = "synsourcemapper"
version = "0.1.0"
requires-python = ">=3.8"

dependencies = [
    "numpy>=1.21.0",
    "torch>=2.0.0",
    "scipy>=1.7.0",
    "matplotlib>=3.4.0",
    "tqdm>=4.62.0",
    "PyYAML>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"